backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
pymupdf==1.24.1
python-dotenv==1.0.0
groq==0.7.0
PyJWT==2.8.0
python-docx==0.8.11
